NoticeTemplateReporter: Add an option for treating projects like packages
ORT has two different notice reporter templates. The notice summary
template just lists all licenses in one large list. The default
template has one joint list for the licenses of all projects and
separate lists for the licenses of each package.

In some cases it makes sense to also list the licenses separetly for (some
of the) projects. In particular, listing the licenses for
SpdxDocumentFile projects separately is interresting, because this
allows to create a notice by package report also in cases where no
package manager is involved and the source code of the project dependencies
is just included in the source tree of the analyzed project.

In order to support that use case the scan results have to be
deduplicated for the projects which are to be reported separately in the
notices, because the scan results of projects may overlap - which would
in turn result in licenses being reported twice, in the list for the
respective package and in the joint licenses list for the projects.

Signed-off-by: Frank Viernau <frank.viernau@here.com>